Extract ItemProfilesManager
