+++ common/src/main/java/com/wynntils/wynn/item/IngredientItemStack.java
@@ -4,8 +4,8 @@
  */
 package com.wynntils.wynn.item;
 
-import com.wynntils.core.webapi.WebManager;
 import com.wynntils.wynn.item.parsers.WynnItemMatchers;
+import com.wynntils.wynn.model.ItemProfilesManager;
 import com.wynntils.wynn.objects.profiles.ingredient.IngredientIdentificationContainer;
 import com.wynntils.wynn.objects.profiles.ingredient.IngredientItemModifiers;
 import com.wynntils.wynn.objects.profiles.ingredient.IngredientModifiers;
@@ -39,7 +39,7 @@ public class IngredientItemStack extends WynnItemStack {
             throw new IllegalStateException("Matcher did not match for IngredientItemStack");
         }
 
-        ingredientProfile = WebManager.getIngredients().get(matcher.group(1));
+        ingredientProfile = ItemProfilesManager.getIngredients().get(matcher.group(1));
 
         isGuideStack = false;
         guideTooltip = List.of();
